RollingGoalPSoC test og optimering af sensor - og USBUART - test del

+++ Software/RollingGoalPSoC/Rolling_Goal/RollingGoal.cydsn/sensor.h
@@ -16,10 +16,10 @@
 
 struct sample
 {
-    int32 avg;
+    float avg;
     float rms;
-    int32 min;
-    int32 max;
+    float min;
+    float max;
 };
 
 struct data
